assimilate
